Fix content-disposition handling
This should resolve #21 which should also solve
https://github.com/ecmwf/ecmwf-opendata/issues/54.

The parsing function returns a dictionary, not a tuple. If we
tuple-unpack a dictionary, that will give us the keys of the *first
two* dictionary entries. If there are more than two entries it would
raise a ValueError (too many values to unpack).